Admin system and MP fixes

+++ Content/Tools/Gameplay/Time.cs
@@ -53,7 +53,7 @@ namespace DragonLens.Content.Tools.Gameplay
 
 		public override void RecievePacket(BinaryReader reader, int sender)
 		{
-			Main.time = reader.ReadInt32();
+			Main.time = reader.ReadDouble();
 			Main.dayTime = reader.ReadBoolean();
 			TimePauseSystem.savedTime = reader.ReadInt32();
 			Main.moonPhase = reader.ReadInt32();
